% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Dire Bonjour}
\usage{
dire_bonjour(prenom)
}
\arguments{
\item{prenom}{character to identify person you would like to say hello to}
}
\value{
string with embedded prenom
}
\description{
Function to say hello based on name provided as input
}
